+++ src/controller/getConfiguration.js
@@ -36,6 +36,7 @@ function getUserSettings() {
         disableJavascript: conf.get('disableJavascript'),
         disableJavascriptReact: conf.get('disableJavascriptReact'),
         disableTypescript: conf.get('disableTypescript'),
+        openMinifiedDocument: conf.get('openMinifiedDocument'),
     }
 }